+++ database/migrations/2019_10_01_204740_create_profile_skill_pivot_table.php
@@ -13,9 +13,9 @@ class CreateProfileSkillPivotTable extends Migration
     public function up()
     {
         Schema::create('profile_skill', function (Blueprint $table) {
-            $table->integer('profile_id')->unsigned()->index();
+            $table->bigInteger('profile_id')->unsigned()->index();
             $table->foreign('profile_id')->references('id')->on('profiles')->onDelete('cascade');
-            $table->integer('skill_id')->unsigned()->index();
+            $table->bigInteger('skill_id')->unsigned()->index();
             $table->foreign('skill_id')->references('id')->on('skills')->onDelete('cascade');
             $table->primary(['profile_id', 'skill_id']);
         });
